requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
altair==4.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.1.0
audio-recorder-streamlit==0.0.10
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
emoji==1.6.3
entrypoints==0.4
exceptiongroup==1.3.0
faiss-cpu==1.9.0.post1
fastapi==0.115.14
ffmpeg-python==0.2.0
ffmpy==0.6.0
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2024.12.0
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gradio==5.34.2
gradio_client==1.10.3
greenlet==3.1.1
groovy==0.1.2
grpcio==1.70.0
h11==0.16.0
h5py==3.12.1
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.24.1
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
# nvidia-cublas-cu12==12.4.5.8
# nvidia-cuda-cupti-cu12==12.4.127
# nvidia-cuda-nvrtc-cu12==12.4.127
# nvidia-cuda-runtime-cu12==12.4.127
# nvidia-cudnn-cu12==9.1.0.70
# nvidia-cufft-cu12==11.2.1.3
# nvidia-curand-cu12==10.3.5.147
# nvidia-cusolver-cu12==11.6.1.9
# nvidia-cusparse-cu12==12.3.1.170
# nvidia-cusparselt-cu12==0.6.2
# nvidia-nccl-cu12==2.21.5
# nvidia-nvjitlink-cu12==12.4.127
# nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psycopg2==2.9.10
pyarrow==19.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.12.1
safehttpx==0.1.6
safetensors==0.5.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.37
starlette==0.46.2
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
text2emotion==0.0.5
tf_keras==2.18.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.3
toolz==1.0.0
torch==2.6.0
torchaudio==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.1
triton==3.2.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.3
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2


# previously-installed
# streamlit
# pandas
# python-dotenv
# psycopg2
# tf-keras
# emoji==1.6.3
# huggingface_hub
# torch 
# sqlalchemy
# tensorflow
# transformers
# speechrecognition
# audio-recorder-streamlit
# nltk
# text2emotion


# absl-py==2.1.0
# altair==4.2.2
# astunparse==1.6.3
# attrs==25.1.0
# audio-recorder-streamlit==0.0.10
# blinker==1.9.0
# cachetools==5.5.2
# certifi==2025.1.31
# charset-normalizer==3.4.1
# click==8.1.8
# colorama==0.4.6
# emoji==1.6.3
# entrypoints==0.4
# faiss-cpu==1.10.0
# filelock==3.17.0
# flatbuffers==25.2.10
# fsspec==2025.2.0
# gast==0.6.0
# gitdb==4.0.12
# GitPython==3.1.44
# google-pasta==0.2.0
# greenlet==3.1.1
# grpcio==1.70.0
# h5py==3.13.0
# huggingface-hub==0.29.1
# idna==3.10
# Jinja2==3.1.5
# joblib==1.4.2
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
# keras==3.8.0
# libclang==18.1.1
# Markdown==3.7
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# mdurl==0.1.2
# ml-dtypes==0.4.1
# mpmath==1.3.0
# namex==0.0.8
# networkx==3.4.2
# nltk==3.9.1
# numpy==2.0.2
# opt_einsum==3.4.0
# optree==0.14.0
# packaging==24.2
# pandas==2.2.3
# pillow==11.1.0
# protobuf==5.29.3
# psycopg2==2.9.10
# pyarrow==19.0.1
# pydeck==0.9.1
# pydub==0.25.1
# Pygments==2.19.1
# python-dateutil==2.9.0.post0
# python-dotenv==1.0.1
# pytz==2025.1
# PyYAML==6.0.2
# referencing==0.36.2
# regex==2024.11.6
# requests==2.32.3
# rich==13.9.4
# rpds-py==0.23.1
# safetensors==0.5.3
# six==1.17.0
# smmap==5.0.2
# SpeechRecognition==3.14.1
# SQLAlchemy==2.0.38
# streamlit==1.42.2
# streamlit_chat_widget==0.2.0
# sympy==1.13.1
# tenacity==9.0.0
# tensorboard==2.18.0
# tensorboard-data-server==0.7.2
# tensorflow==2.18.0
# tensorflow-io-gcs-filesystem==0.31.0
# tensorflow_intel==2.18.0
# termcolor==2.5.0
# text2emotion==0.0.5
# tf_keras==2.18.0
# tokenizers==0.21.0
# toml==0.10.2
# toolz==1.0.0
# torch==2.6.0
# tornado==6.4.2
# tqdm==4.67.1
# transformers==4.49.0
# typing_extensions==4.12.2
# tzdata==2025.1
# urllib3==2.3.0
# watchdog==6.0.0
# Werkzeug==3.1.3
# wrapt==1.17.2